refactor(history): decompose HistoryCard into history-card subcomponents
Extract HistoryCard's 910+ lines into a focused component composition layer
and introduce eight functional subcomponents in src/components/history/history-card/:
- HistoryCardHeader: timestamp, source, tags, transfer progress
- HistoryCardContent: polymorphic content preview (text/code/link/image/file)
- HistoryCardActions: copy/delete/favorite buttons
- HistoryCardTransferProgress: linear progress bar for file operations
- FileGlyph: file extension badge with color coding
- *EntryContent: type-specific preview renderers
- Utility modules: history-card-utils, file-entry-utils, file-glyph-utils,
  useResourceImageUrl

Improves testability (extract content rendering) and maintainability
(smaller, focused components). Add test cases for file cards and code
preview line numbers.
