+++ src/components/history/history-card/file-entry-utils.ts
@@ -1,0 +1,6 @@
+import type { ClipboardFileItem } from '@/lib/clipboard-entry'
+import { isImageFileName } from '@/lib/clipboard-utils'
+
+export function isSingleImageFile(item: ClipboardFileItem): boolean {
+  return item.file_names.length === 1 && isImageFileName(item.file_names[0] ?? '')
+}
