+++ src/components/history/history-card/file-glyph-utils.ts
@@ -1,0 +1,3 @@
+export function getFileExtLabel(name: string): string {
+  return name.split('.').pop()?.toUpperCase() || 'FILE'
+}
